+++ back/src/main/java/board/pjt/back/dto/category/CategorySelectResponseDto.java
@@ -1,22 +1,22 @@
 package board.pjt.back.dto.category;
 
 public class CategorySelectResponseDto {
-    private Integer category_id;
+    private long category_id;
     private String name;
 
     public CategorySelectResponseDto() {
     }
 
-    public CategorySelectResponseDto(Integer category_id, String name) {
+    public CategorySelectResponseDto(long category_id, String name) {
         this.category_id = category_id;
         this.name = name;
     }
 
-    public Integer getCategory_id() {
+    public long getCategory_id() {
         return category_id;
     }
 
-    public void setCategory_id(Integer category_id) {
+    public void setCategory_id(long category_id) {
         this.category_id = category_id;
     }
