+++ back/src/main/java/board/pjt/back/dto/friendRequest/GetSenderEmailByFriendRequestIdRequestDto.java
@@ -3,12 +3,12 @@ package board.pjt.back.dto.friendRequest;
 import java.util.Objects;
 
 public class GetSenderEmailByFriendRequestIdRequestDto {
-    private int friend_request_id;
+    private long friend_request_id;
 
     public GetSenderEmailByFriendRequestIdRequestDto() {
     }
 
-    public GetSenderEmailByFriendRequestIdRequestDto(int friend_request_id) {
+    public GetSenderEmailByFriendRequestIdRequestDto(long friend_request_id) {
         this.friend_request_id = friend_request_id;
     }
 
@@ -32,11 +32,11 @@ public class GetSenderEmailByFriendRequestIdRequestDto {
                 '}';
     }
 
-    public int getFriend_request_id() {
+    public long getFriend_request_id() {
         return friend_request_id;
     }
 
-    public void setFriend_request_id(int friend_request_id) {
+    public void setFriend_request_id(long friend_request_id) {
         this.friend_request_id = friend_request_id;
     }
 }
